refactor: 💡 Removed "Sort Sets" button and made HiddenSets
